requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.0
pandas==2.0.3
networkx==3.1
numpy==1.24.3
requests==2.31.0
redis==5.0.0
python-dateutil==2.8.2
pyarrow==13.0.0
